requirements.txt
# Core
Django==4.2.10
asgiref==3.7.2
sqlparse==0.4.4

# GraphQL
graphene-django==3.2.2
graphql-core==3.2.3
django-graphql-jwt==0.3.4
graphene-file-upload==1.3.0

# Auth / API utils
django-filter==23.5
django-cors-headers==4.4.0

# Media
Pillow==10.2.0

# DB
psycopg2-binary==2.9.9

# Env & server
python-dotenv==1.0.1
gunicorn==21.2.0
